+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
-25/11/2020
+28/11/2020
+- Add @connections to Trade DynamoDB table (400 KB items size limit)
 - Added S3 API for JSON file uploads.